Removed unneccesary files in helper folder | Used isSuperAdmin to authenticate users
