--- rubrics-ui/src/app/api/helpers/academy/academyObjectTypes.ts
@@ -1,6 +1,0 @@
-export enum AcademyObjectTypes {
-  bytes = 'bytes',
-  guides = 'guides',
-  timelines = 'timelines',
-  simulations = 'simulations',
-}
