--- rubrics-ui/src/app/api/helpers/constants/constants.ts
@@ -1,6 +1,0 @@
-export const TOP_CRYPTO_PROJECTS_SPACE_ID = 'top-crypto-projects';
-
-export const PredefinedSpaces = {
-  CRYPTO_GELATO: 'top-crypto-projects',
-  TIDBITS_HUB: 'tidbitshub',
-};
