--- rubrics-ui/src/app/api/helpers/deprecatedSchemas/models/course/CourseTopics.ts
@@ -1,14 +1,0 @@
-import { TopicQuestionModel } from './TopicQuestionModel';
-import { TopicReadingModel } from './TopicReadingModel';
-import { TopicSummaryModel } from './TopicSummaryModel';
-import { TopicExplanationModel } from './TopicExplanationModel';
-
-export interface GitCourseTopicModel {
-  title: string;
-  key: string;
-  details: string;
-  explanations?: TopicExplanationModel[];
-  questions?: TopicQuestionModel[];
-  readings?: TopicReadingModel[];
-  summaries?: TopicSummaryModel[];
-}
