--- rubrics-ui/src/app/api/helpers/deprecatedSchemas/models/course/TopicConfig.ts
@@ -1,4 +1,0 @@
-export interface TopicConfig {
-  showExplanations: boolean;
-  showHints: boolean;
-}
