--- rubrics-ui/src/app/api/helpers/deprecatedSchemas/models/course/TopicExplanationModel.ts
@@ -1,6 +1,0 @@
-export interface TopicExplanationModel {
-  shortTitle: string;
-  title: string;
-  key: string;
-  details: string;
-}
