--- rubrics-ui/src/app/api/helpers/deprecatedSchemas/models/course/TopicQuestionModel.ts
@@ -1,16 +1,0 @@
-export interface QuestionChoice {
-  content: string;
-  key: string;
-}
-
-export interface TopicQuestionModel {
-  uuid: string;
-  type: string;
-  content: string;
-  hint: string;
-  explanation: string;
-  answerKeys: string[];
-  subTopics: string[];
-  difficultyLevel: string;
-  choices: QuestionChoice[];
-}
