--- rubrics-ui/src/app/api/helpers/deprecatedSchemas/models/course/TopicReadingModel.ts
@@ -1,9 +1,0 @@
-export interface TopicReadingModel {
-  uuid: string;
-  title: string;
-  details: string;
-  type: string;
-  url: string;
-  shortTitle: string;
-  subTopics: string[];
-}
