--- rubrics-ui/src/app/api/helpers/deprecatedSchemas/models/course/TopicSummaryModel.ts
@@ -1,6 +1,0 @@
-export interface TopicSummaryModel {
-  title: string;
-  key: string;
-  details: string;
-  shortTitle: string;
-}
