--- rubrics-ui/src/app/api/helpers/deprecatedSchemas/models/timeline/TimelineEvent.ts
@@ -1,8 +1,0 @@
-export interface TimelineEvent {
-  title: string;
-  uuid: string;
-  date: string;
-  summary: string;
-  fullDetails?: string | null;
-  moreLink?: string | null;
-}
