--- rubrics-ui/src/app/api/helpers/deprecatedSchemas/models/timeline/TimelineModel.ts
@@ -1,15 +1,0 @@
-import { TimelineEvent } from './TimelineEvent';
-
-export interface TimelineModel {
-  id: string;
-  name: string;
-  excerpt: string;
-  content: string;
-  thumbnail?: string;
-  created: string;
-  publishStatus: string;
-  events: TimelineEvent[];
-  admins: string[];
-  tags: string[];
-  priority: number;
-}
