--- rubrics-ui/src/app/api/helpers/getAuthor.ts
@@ -1,6 +1,0 @@
-export function getAuthor(accountId: string) {
-  return {
-    name: `Account - ${accountId}`,
-    email: '',
-  };
-}
