--- rubrics-ui/src/app/api/helpers/git/ensureDirectoryExistence.ts
@@ -1,11 +1,0 @@
-import fs from 'fs';
-import path from 'path';
-
-export function ensureDirectoryExistence(filePath: string) {
-  const dirname = path.dirname(filePath);
-  if (fs.existsSync(dirname)) {
-    return true;
-  }
-  ensureDirectoryExistence(dirname);
-  fs.mkdirSync(dirname);
-}
