--- rubrics-ui/src/app/api/helpers/gitGuides/model/GitGuideIntegrations.ts
@@ -1,9 +1,0 @@
-export interface GitGuideIntegrations {
-  discordRoleIds: string[];
-  // number of answers that should be correct in order to get discord score
-  discordRolePassingCount?: number;
-  discordWebhook?: string;
-  projectGalaxyCredentialId?: string;
-  projectGalaxyOatMintUrl?: string;
-  projectGalaxyOatPassingCount?: number;
-}
