--- rubrics-ui/src/app/api/helpers/gitGuides/model/GitGuideStep.ts
@@ -1,10 +1,0 @@
-import { GitUserDiscordConnect } from '@/app/api/helpers/gitGuides/model/GitGuideStepItem';
-import { GitGuideQuestion } from './GitGuideQuestion';
-import { GitUserInput } from './GitUserInput';
-
-export interface GitGuideStep {
-  content: string;
-  name: string;
-  stepItems: (GitGuideQuestion | GitUserInput | GitUserDiscordConnect)[];
-  uuid: string;
-}
