--- rubrics-ui/src/app/api/helpers/gitGuides/model/GitUserInput.ts
@@ -1,6 +1,0 @@
-import { GitGuideStepItem } from './GitGuideStepItem';
-
-export interface GitUserInput extends GitGuideStepItem {
-  label: string;
-  required: boolean;
-}
