--- rubrics-ui/src/app/api/helpers/types/discord.ts
@@ -1,6 +1,0 @@
-export interface DiscordAuthor {
-  name: string;
-  url?: string;
-  icon_url?: string;
-  proxy_icon_url?: string;
-}
