+++ rubrics-ui/src/app/edit-program/page.tsx
@@ -1,27 +1,10 @@
-'use client';
-import React, { useEffect, useState } from 'react';
-import ProgramEdit from '@/components/ProgramEdit/ProgramEdit';
-import { getSession } from 'next-auth/react';
-import { useLoginModalContext } from '@dodao/web-core/ui/contexts/LoginModalContext';
-import { SessionProps } from '@/types/rubricsTypes/types';
+import ProgramEdit from '@/components/EditProgram/EditProgram';
 function EditProgram() {
-  const { showLoginModal, setShowLoginModal } = useLoginModalContext();
-  const [session, setSession] = useState<SessionProps | null>(null);
-  const isUserLoggedIn = async () => {
-    const session = await getSession();
-    setSession(session as SessionProps | null);
-    if (!session) {
-      setShowLoginModal(true);
-    } else {
-      setShowLoginModal(false);
-    }
-  };
-
-  useEffect(() => {
-    isUserLoggedIn();
-  }, [showLoginModal]);
-
-  return <div>{session && <ProgramEdit />}</div>;
+  return (
+    <div>
+      <ProgramEdit />
+    </div>
+  );
 }
 
 export default EditProgram;
